requirements.txt
datetime
lxml
python-dotenv
requests
yt-dlp